prev CRUD app works properly. change PK long for User table

+++ src/main/java/com/example/SpringBoot_Security/DAO/UserDAO.java
@@ -11,9 +11,9 @@ public interface UserDAO {
 
     void save(User user);
 
-    User getOneUser(int id);
+    User getOneUser(Long id);
 
     public void updateUser(User user);
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
 }
